tyle file for research Policy.
See also discussion at

http://forums.zotero.org/discussion/13748/style-request-research-policy

My read of the Research Policy guidelines:

http://www.elsevier.com/wps/find/journaldescription.cws_home/505598/authorinstructions#56000

(scroll down to References)

is that Research Policy is Elsevier Harvard but with a few changes:
- no abbreviated journal titles (remove form="short" from container (in else group)
- making use of the p. 45 style locator (which the Elsevier Harvard doesn't?)
- adjusting scope of name disambiguation.
- issue numbers only need to be given if each issue of that journal begins its numbering from page 1.)
- page-range-format
- punctuation in bibliography

Scope of name disambiguation is mentioned in the guide, but doesn't completely solve the question, "(family name only, although include initials if the paper refers to work by two different authors with the same family name)"

Looking at a few papers it seems to be handled pretty idiosyncratically.  For example this paper (by a C Freeman, citing many papers by C Freeman):

Freeman, C. and Soete, L. (2009). Developing science, technology and innovation indicators: What we can learn from the past. Research Policy, 38(4):583 – 589.

cites a publication by an R Freeman in text as Richard Freeman, but the C Freeman articles as just Freeman.  In other places the same author is sometimes given two initials in the bibliography, sometimes one but just the last name in the in-text citation.

My take is that this journal minimizes initials in citations. the rule above specifically says papers with two authors with same last name should have initials and therefore I think it should use  givenname-disambiguation-rule="by-cite" and while it asks for initials, if those match then full given names (as with the Richard Freeman above). disambiguate-add-givenname="true" givenname-disambiguation-rule="by-cite"

page range format.  One citation (Pavitt) in the guidelines seem to suggest collapsing repeated or chicago, but looking at actual articles this doesn't appear to be done.  Going to leave it off for now.

Issue numbers, from guidelines "In addition, issue numbers only need to be given if each issue of that journal begins its numbering from page 1.)" Not sure if this can be done in any automated way.  Ignoring.

Punctuation in bibliography.  Removed , after title for journal articles.  Removed period after Eds.  Again this is how it is in the style guide, but most actual articles use the period.  Sigh.