Add style for Ergoscience
This adds a unique style for the German journal "Ergoscience" as
published in Schulz-Kirchner Verlag. The style is heavily based on the
Acta Materialia one, but required a few tweaks that I could not find in
any existing unique style. Things like patent citation probably still
look out of place, but the rest should be quite OK.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>