Add legislation to bill in conditional tests, part 1.
The Statute field in Zotero will soon be mapped to legislation.
No styles yet discriminate between enacted laws (legislation) and
proposed law drafts (bill), so it is sufficient to add legislation
to tests where it is missing.

This changeset covers two styles where legislation was sometimes
included in tests, sometimes not.