Add legislation to bill in conditional tests, part 2.
The Statute field in Zotero will soon be mapped to legislation. No
styles yet discriminate between enacted laws (legislation) and
proposed law drafts (bill), so it is sufficient to add legislation to
tests where it is missing.

This changeset covers styles where bill was included in tests,
but legislation was not. There should be no effect on current
style behavior. Styles receiving this change will not be
affected by the new mapping, but it will be open to style
designers to apply separate treatment to bills and enacted
legislation.