Updated Geochimica et Cosmochimica Acta style
The citation style for the Geochimica journal did not match citations in
that publication - specifically et-al use was required for 3 or more
authors, while the full author list should be reserved for the
bibliography. However, the original contributor of the CSL could hardly
be blamed for this oversight as the Geochimica site at [1] provides
details of bibliography style while saying nothing about citation style.

The example Geochimica paper on which this updated style is based is
available from [2] (sorry, not open access).

Please note there is a horrible hack in the updated CSL: the et-al-min
setting specified in <citation> seems to "leak" into <bibliography>, at
least when using the CSL editor at citationstyles.org.

To counter this, a ridiculously large et-al-min setting has been
specified for the bibliography element. I'm not sure yet if this is a
bug, and if so where it might be (csl-editor or citeproc-js) but if I
can track it down I'll file a bug report with the appropriate project.

[1] http://www.elsevier.com/journals/geochimica-et-cosmochimica-acta/0016-7037/guide-for-authors#67000

[2] http://www.sciencedirect.com/science/article/pii/S0016703712006850