Fix information systems research
After the fix c3c85924, the citation-locator is now appended with the
group delimiter " ", ie citations look like this: (2012 pp. 12-14)

Checking articles, they have to look like this: (2012, pp. 12-14).
This can be fixed by taking the citation-locator out of the group that
is delimited with " " and add the citation-locator with the ", " prefix.