Create iso690-author-date-fr-without-abstract.csl
This is the very same style as the original iso690-author-date-fr except I removed abstract and notes from bibliography. The original authors insisted to keep the abstract, as they found it valuable, but I know a lot of people (including me) fed up with removing abstracts from their bibliographies. I thought this was the best way to proceed.