Fix editor name part order in style "Harvard Durham University Business School"
In the style "Harvard Durham University Business School" the editors were in the bibliography written as "Richard E. Nisbett" instead "Nisbett, R.E.". That did not affect the authors, but only editors.

This bug was copied over from the style "Harvard - Imperial College London", on which Durham style is based.