Fix et-al in ecology-letters bibliography section
The "et al" phrase was being printed for all references, even those with
all authors printed. To fix this I have removed a label line and added a
necessary prefix space between the authors and the "et al". Besides
forcing "et al" to appear for all references, I'm not sure if there was
another purpose to the label line. Perhaps there is a better fix.