Remove citation-number block from ecology-letters
The author instructions specify that the bibliography should be:
"listed in alphabetical order at the end of the paper, and numbered
sequentially (in the reference list only). A number and a line break
should be added between each reference."
http://onlinelibrary.wiley.com/journal/10.1111/(ISSN)1461-0248/homepage/ForAuthors.html

The numbers that were added were the `citation-numbers`, which
corresponded to the order cited, not the alphabetical order. It may not
be possible to add this type of numbering in CSL. Further, if these
numbers can be added, there needs to be a line break after the number.
Therefore, perhaps it is better if the number is not added at all and
the user can add these as a last-minute manual edit before paper
submission.

The official Ecology Letters format (as published) omits these numbers.