Create frontiers-medical-journals.csl
The 'Frontiers in' series of journals has two different citation styles and two different bibliography styles.  One set of styles is for journals in the fields of science and engineering, the other set of styles is for journals in the field of medicine.  These two sets of styles are documented at http://www.frontiersin.org/about/AuthorGuidelines#References .

Previous contributors created .csl files for all the 'Frontiers in' series of journals, but they applied science and engineering rules to all of the journals, even the medical journals.  Since I want to publish in Frontiers in Immunology, I was disheartened to find that a new style needed to be created. 

I have created an appropriate .csl file, using http://www.zotero.org/styles/frontiers as my template.  My new file is frontiers-medical-journals.csl.  It is, like its template, designed to be used as a base for other styles.  For now I am using it as my selected style in my document; it seems to work well with Word and Zotero.  After you accept it, I recommend that someone change all the styles for all the 'Frontiers in' medical journals to use this new style as their base.  If you want, I will do that.  (Although if you preferred to have someone else do that, I would be more than happy to let you or someone else take care of that task.)

There remains an interesting design question about the inheritance tree for 'Frontiers in' journals.  In frontiers-medical-journals.csl, I have created a peer of frontiers.csl, that duplicates the old code by copy-and-paste.  Maybe there should be a base for all 'Frontiers in' journals, then two derived styles, frontiers.csl for science and engineering, and my new one for medicine.  This alternative approach would eliminate xml code duplication, but take more work, and add a layer to the hierarchy.  I don't even know if the csl language and infrastructure would allow an additional layer.

I don't know whether I can even test a frontiers-in-medicine.csl as a parent yet until you accept and post it so it, because frontiers-in-immunology would refer to a non-existent URL as a parent.  Therefore I have not tested frontiers-in-medicine.csl as a parent yet.

Note that there is no journal named Frontiers in Medicine; I deliberately chose an unused journal-name for this new parent file.

Chris Forden
cforden at-sigil comcast.net