Create frontiers-in-immunology.csl
This is an update from the existing http://zotero.org/styles/frontiers-in-immunology ; I have switched the parent to the new frontiers-in-medicine.csl because this and many other 'Frontiers-in' journals take styles different from 'Frontiers-in' science and engineering journals.  I made this change manually, but after this is proven OK, I plan to use utilities/generate_dependent_styles/generate_styles_from_data.rb to create updates for the other medical journals in the 'Frontiers in' series.  That will seriously increase the number of files in https://github.com/citation-style-language/styles/ .  But that is the plan, right?