A squashed version of commits for adding the oxford-university-press-humsoc style.
Added a file for style.  I copied apa 5th edition as the starting point, will show changes from there.

Updated metadata

And actually validated the metadata

Trying to get publisher-place: publisher for chapter.  Not quite sure how it's controlled, doesn't seem to be working at present.  Thought that the match="any" with chapter would do it on line 167, but it didn't.

chapter now has publisher: place as called for in guidelines, used a special else-if just for chapter

Removed the comma from the in-text citations.  Used two mdashes for the subsequent author abbreviation (copied out of their pdfs)

removed italic from volume

removed URL unless from webby sources

hanged delimiter in locator between volume/issue: pages.  Hope this is the right way to do that.

Added output of DOI.  They don't say anything about it and I think it's important.

Moved the date part so that it never appears in the field after the names.  I moved it to after the conference title etc.  This may be incorrect but there's nothing to say either way.

Added < and > around the URL field whereever it outputs (macro="URL").  Might need to do that around the archive field too, but I don't really know what that is :)

Trying to follow requirements for submission of CSL.  Fixed tabs for spaces, added template link (and removed authors as directed)

Oxford HUMSOC looks like an abbreviation, but I can't find what it stands for.  Probably Humanities and Social Sciences, but who knows?

Fixed (ed/eds) thing using locale as mentioned in pull request thread.  Also changed et al style to min=3 after=1, which matches what is actually done in the journals, rather than what the "mini-checklist" calls for.

Changed the title to Oxford University Press Humanities and social-sciences.  Perhaps I should be changing the title of the file as well, but since OUP consistently refers to this as HUMSOC (and not as Humanities and Social-Sciences), I think this is a mistake.

Added dependent style for the Research Evaluation journal.

Getting title-short right.

Changed id to oxford-university-press-humsoc.  Changed dependent style as well.

Renamed humsoc csl file to reflect new id