Create ISO-690 (author-date, Revista Ciencias Técnicas Agropecuarias).csl
This style is based on the norm ISO 690: 2010 and it was created for the Journal  Ciencias Técnicas Agropecuarias of the Agrarian University of Havana (UNAH) in Cuba. It cite in the text according to the form author-year and it organizes the bibliography alphabetically. It presents as distinctive features that in the bibliography the authors are presented in the short form, after the authors two points are placed and the year always appears at the end. It is able to recognize, to organize and to fill a very wide quantity of articles correctly for what can be used by any user.