chicago full note tr please look in. important
Hi The number of Authorized, Chicago Full Note TR footnotes giving the classic style procedure to be used in Turkey is the Turkish version. Arrangements were made according to the rules of Turkish language. Location of the labels have been changed. This is an initial step to the next work will be. This study will be presented at the symposium, Bayburt University as inappropriate. Chicago manual of style is also not enough to work.Chicago manual of style is also not enough to work.