Create APOLLON_Hochschule_der_Gesundheitswirtschaft_(German).csl
APOLLON-Hochschule-der-Gesundheitswirtschaft-(German).csl
The existing style "American Journal of Physical Anthropology" was edited to "APOLLON Hochschule der Gesundheitswirtschaft (German)" primarily for authors of the private College.
Several changes were made:
The authors names are delimited by a semicolon, the first names are initialized with a dot, sort seperator is a comma. The date is in round braces. The title is italic. The language was changed to German. In case of a Journal the volume and issue (in braces) and the exact issues date was added. The container title seperates with a comma to the issued date. The access date of URLs was added. The term "et al" is used when more than 4 authors are present. The inline citation uses "et al" when 2 or more authors are present. The editor is in braces and seperates with a colon.