Update irish-historical-studies.csl
This revision suppresses the access date when it is not required for IHS style.

The revision simplifies the citation of e.g. paper journal articles also available online for which an online access date has been scraped automatically. Currently for such items the access date must be manually removed from the library entry for the style to fomat correctly.