Fixed issue with inconsistent volume identifiers.
This fixes and issue with the way that the volume number is displayed in first footnoted citations, for clarity and uniformity with subsequent citations.  It takes the volume number out of the parentheses, and formats it as xx:xxx-xxx.  The bibliography has not changed, but due to previous formatting, a new macro had to be added.