Create zeitschrift-fuer-deutsche-philologie.csl
This should fulfill this <a href="https://forums.zotero.org/discussion/21423/style-request-zeitschrift-fuer-deutsche-philologie-zfph-german-humanities/">old style request</a>. Please note, I suggest in this style to use encyclopediaArticle for "Werke in Editionen" and the style will then print the "Ders./Dies." instead of the editors (which are then the same as the authors). I will ignore gender-specific wording here.