Create spectroscopy-letters
I've started working from the applied spectroscopy style as mentioned in the summary. 
There are still a few issues, but at least it seems to be compliant to 100% with journal references
Issue: a comma after the date still appears in the book references and I didn't know how to remove it. Probably it works with conditionals...)