cmslhi.csl
This is the code for use with the referencing style used at the Iceland Academy of the Arts, based on the Chigaco Manual of Style, 16th ed. (full note). BA students writing in Icelandic use this style within the Zotero application. MA students use the original style of referencing as it appears on the Zotero site and in application.