Update journal-of-neurochemistry.csl
There is a problem with validation. I must admit that I do not know how to solve this problem. If I want in citation enter term et al. with italics, I believe that I can now save it in only one way - by adding an entry <et-al font-style = "italic" prefix = "" />, but this line is reported as an error.