Create journal-of-retailing.csl
I checked current articles which are indeed looking like the same style as the Journal of Interactive Marketing except the rendering of the issue number ( https://forums.zotero.org/discussion/39790/style-request-journal-of-interactive-marketing/#Item_32 ). However, the citation style guide which is under the documentation link looks differently in some details (not italics, order of editors for chapter and conference paper, comma/point after author names and year). @grnathan2000 : It is an Elsevier journal; do you know anything about this citation style and/or the style guide?