Update and rename universita-pontificia-salesiana-en.csl to universita-pontificia-salesiana.csl
The dependent styles (universita-pontificia-salesiana-fr/es/pt-br) are not working (they use the English locale). I think the reason is that the parent has the default locale "en-US".
I'm proposing a new generic UPS style, without default locale, and will correct the "independent-link" link later).
Please, suggest another change if this is not the reason of the problem.