Create modern-language-association-annotated-bib-with-url.csl
This was developed for the International Writing Centers Association's Antiracism SIG annotation project. It's based on the latest MLA .csl file with the addition of URLs in opening and closing brackets and the use of the abstract field for annotations. It aligns with MLA guidelines for annotated bibliographies, so it could be used by students for annotated bibliographies.