Update rapid-communications-in-mass-spectrometry.csl
This should improve the correct punctation in the style. I don't have any examples for `patent`, `thesis`, or `webpage`, but it seemed natural to fix it in this way. However, in these cases the `title` is renderd twice (which seems like an error) and I don't know which version is the correct one. Ping @hanvanzan

(I excluded this style from #1334, because it is too special)