Update wiley-vch-books.csl
Unfortunately, the previous version was not conforming the instructions on citation style by Wiley. As it is not based on any template now, it supports articles, books and chapters well, everything else uses fallback style, which is based on article. But I have no examples to check how it works. On the other hand, it supports the most important types of references.