Fix for items with both title and reviewed-title
Adapted from the following example for the notes style (CMoS, 16e, 14.215 – there seem to be no such examples in ch. 15):
2. David Kamp, “Deconstructing Dinner,” review of *The Omnivore’s Dilemma: A Natural History of Four Meals*, by Michael Pollan, *New York Times*, April 23, 2006, Sunday Book Review, http://www.nytimes.com/2006/04/23/books/review/23kamp.html.
Logic: If there's a reviewed-author and a title only, assume the title is the title of the reviewed work. If there's a reviewed-author, a title, and a reviewed-title, assume the title is the title of the review, and reviewed-title is the title of the reviewed work.
This would of course work better if Zotero had a proper review item type, and if CSL had reviewed-editor, reviewed-director, etc., and supported entering more than one reviewed work, too.