Update british-ecological-society.csl
There were lots of brackets and commas for the editors in an edited book. I tried to fix it (lines 136-142); hopefully I did not screw up with groups!
There was also a problem with the prefix and suffix of collection title (line 133). It was not separated from book title (so I added a comma as prefix) and it ended with a period although the editor list comes right after in brackets.