[apa.cls] Add possibility to cite data sets
This is described here http://blog.apastyle.org/apastyle/2013/12/how-to-cite-a-data-set-in-apa-style.html . However, we will use the newest recommondation about how to format doi's here as well. I added also the case, that it is possible to have data with a title contained in a data set with a title as outlined here https://www.zotero.org/support/dev/translators/datasets

Let me know what you think.