Update nature.csl
This is the example from the Nature guidelines for citing a website:
"Buckley, M. and Reid, A., Global food safety - keeping food safe from farm to table. Technical report. (2010) Available at: insert website here. (Accessed: 4th November 2012)".

I have therefore edited lines 45-52 to format the URL and add the access date, and added lines 119-123 so that URL and access date are shown only for webpages. It works as intended (as far as I could check), but I'm not sure whether it is the best way to do so!