Create HSU-ENGR.csl
A modification of the ASCE citation style that conforms to the style guidelines of the Humboldt State University Engineering Department. It was created for use in ENGR 492 Spring 2016 but is available for anyone. This style differs from the existing ASCE citation style by adding a comma between the author and date in in-text citations.