Create herpetological-conservation-and-biology.csl
This is a newly created CSL style for the journal Herpetological Conservation and Biology.  Please do not yet post to the repository as it will require final approve from our Board of Directors before being made publicly available.  I am posting here as a pull request so that it can be preliminarily reviewed for technical correctness.