style for the "Journal fuer Kunstgeschichte"
This is a style for the "Journal fuer Kunstgeschichte" / "Journal of Art History". I have written this style a few months ago and you requested me to share it. Here is the referring discussion in the Zotero forum: https://forums.zotero.org/discussion/55979/stil-infoclio-anpassen/ 
Unfortunately there is no online documentation for this style. I have used my own copies in order to create it. Note that the field "short title" must be edited according to the instructions in "summary" if you want it to work properly.
I have never created a new file. I hope this works.