Spell out "editor(s)" and add comma between and (#2050)
* Spell out "editor(s)" and add comma between and

Following example of pg. 20 in the section `Documents having compilers/editors` of http://files.geology.utah.gov/online/c/c-105.pdf. Fixes the abbreviation `eds.` to the correct `editors` and adds a comma between the initials of the first editor and the word "and".

* Addressing note suggestions by @rmzelle

Fixes incorrect name `initialize-with` of `.,` and removed verbose use of `label form="long"`

* Add space for editor

I apologize. I did test it in my zotero script editor, but i guess I missed the space with my changes in github.  I will copy and paste the github scripts into a preview tool in the future.