Update idojaras-quarterly-journal-of-the-hungarian-meteorological-service.csl
Thank you adam3smith for the review.

All suggestions corrected except the one:
"my sense is that you'd probably want to get rid of all the collection titles. They're not that commonly used, and series (which is what they map to) aren't mentioned in the journal guide". 

The series title (in: ) was added to the style based on my previous experience with the journal, as this is "the usual way they do", although it is not explicitly defined in the guidelines. Therefore, I'd prefer to keep that part.