Create negative-results.csl
We are pleased to introduce you to Negative Results, a journal that publishes articles that challenge current knowledge, existing models, and tenets. The articles are based on rigorous, and well-documented results that do not support these models or even refute them. Scientists usually communicate their findings by publishing results in specialized peer-reviewed journals. However negative findings – those that do not agree with what the researchers hypothesized – are often disregarded, neglected and therefore not put forward for publication. The stigma surrounding negative findings means that they are a low priority for publication.