Create uspekhi-gerontologii (Rusian)
This is a style for the Russian journal "Успехи геронтологии" (uspekhi-gerontologii) -- Advances in Gerontology.
There is a problem. I cannot realise how to make it to write "et al." for English or Russian items differently. It should be "et al." for English and "и др." for Russian items. Could you help me to do this.