Create studia-historica-upsaliensa.csl
Style for student essays and dissertation series of the department of History, Uppsala University. Formats item type "Manuscript" and any item type with information in BOTH of the fields "Archive" and "Location In Archive" separately in the bibliography, in order to facilitate the use of Zotero (and others) to cite archival sources.