Update deutsches-archaologisches-institut.csl
Added display of booktitles for chapters. Changed 'et al.' to 'u. a.'. Stripped periods from title fields. Display of '(Hrsg.)' instead of '(eds.)'. Only display edition if number and superscripted. No commas before page numbers, except for articles. Included series number (and series title) for publications in series.
One problem left: In chapters the editors of the book are not displayed correctly; e.g. it should be 'M. Alexianu' and not 'Alexianu, Marius'