MHRA fallback for non-numeric volume
Print-out the volume field "as is" when it is not numeric (e.g. if the user has typed in the roman numerals).
See https://forums.zotero.org/discussion/63174/problem-with-modern-humanities-research-association-3rd-edition#latest

Side note: the Zotero Style Editor seems to ignore the "small-caps" font-variant setting for cs:text, but it works fine when inserted in a Word document.