Create journal-of-sport-and-health-science.csl
This responds to issue https://github.com/citation-style-language/styles/issues/1720

After wasting some time with inadequate templates, I realized that 'For further details you are referred to “Uniform Requirements for Manuscripts submitted to Biomedical Journals”' points to Vancouver as fallback (correct me if I am wrong), so I went for a modified Vancouver. The proposed style (1) matches the output for the examples provided in the style guide and (2) includes the tweaks mentioned in https://forums.zotero.org/discussion/37949/style-request-journal-of-sport-and-health-science/.

As always, I will be happy to implement any corrections or improvements.