create veterinary-pathology.csl
Hi,
Had a go at this style. Took me the longest so far as this style is all about macros with lots of conditional layers, so it took me a lot of brain to figure it out. Potentially also made a mess.
Based on Vancouver (Foot note).
Request on Zotero Forum: https://forums.zotero.org/discussion/59276/style-request-veterinary-pathology/p1