Create tropical-animal-health-and-production.csl
Code was here, but had to be modified: https://forums.zotero.org/discussion/42877/style-request-tropical-animal-health-and-production/p1
I am struggling with the chapters. How can I get it to display the page number at the very end? When I add it into the "publisher" macro it duplicates it for normal article-journals, too. 
The rest all works.