Create harvard-university-for-the-creative-arts(figures-and-illustrations).csl
Hi
I have a valid CSL file for University for the Creative Arts which covers text sources in the Harvard style. As an creative arts university students are expected to submit a list of illustrations with each essay, these are in a numbered format. As a consequence I have created an additional style to cover images, I had seen that this was possible as there is a Chicago style which deals with figures and illustrations. I used that style as a starting point and have amended it to cover the Artwork and Maps styles as this will suit our needs. 

If I need to add anything or if there are any errors let me know

Ian