Create scientific-review-engineering-and-environmental-sciences.csl
Style based on "APA no ampersand" with this changes:
1. "et al." from three authors
2. italicise with "." (without space between italicised letters)
The guide for authors only states "use APA style", but the comparison of published articles with outputs from APA style leads to changes of APA style described above.  Unfortunately, available papers on web pages of journal do not use the same format for volume and issue; so the style template for SREES does not change this.