Create journal-of-computer-applications-in-archeology.csl
VIa https://forums.zotero.org/discussion/65737/style-request-journal-of-computer-applications-in-archaeology-caa
As pointed out in the thread, the style is very similar to citizen-science-theory-and-practice as they are both on Ubiquity press.
This journal is new and has now published anything. One to watch. I can imagine the missing dots are mistakes and we can make this style a dependent at some stage.