Prioritize URL over archive name/number
URL is used much more commonly in practice, even for items in databases
with accession numbers (e.g., ProQuest). Archive name/number is used
for reports and theses. Both URL and archive are given as acceptable in
the 2012 Guide to Electronic References (URL is listed first), and URL
seems more reader-friendly.