Vancouver-Elsevier-Spain
Hi, I would like to create this style, it corresponds to many journals in the Spanish portfolio from Elsevier. It is a Vancouver style with some adjustments that the journals have been using for a while.
It relates to the following journals (among others):
-	Medicina clinica: 0025-7753 
-	Endocrinologia, Diabetes y Nutricion: 2530-0164  
-	Gastroeneterologia y Hepatologia: 0210-5705 
-	Cirugia Española: 0009-739X  
-	Revista de Psiquiatria y Salud Mental: 1888-9891   
-	Atencion Primaria: 0212-6567  
-	Educación Medica 1575-1813  
-      Medicina Intensiva  0210-5691 

Please let me know when it will be created.