Properly support personal communications in APA
APA specifies that personal communications should be cited with a
unique in-text style and should be omitted from the bibliography. This
change supports that.

The in-text format is:
(J. P. Campbell, personal communication, August 16, 2010)

To get the “personal communication” term, I applied a locale to
“interview”, which fits conceptually and is not otherwise used in APA.
This could be changed if a “personal-communication” term is added in a
future CSL update (which has been suggested). If there is a better way
to do that, let me know.

If this works, I can make parallel changes to the other APA styles.