Update iica-catie.csl
Have a look at this one.
I've put the "en linea" stuff in the title macro. It now checks if it has a URL and if it is NOT a webpage or post-weblog and will then include the term "online" after the title. it does that for any item type. For webpages/post-weblogs it will add the longer term I had to hardcode in. you can see that in the access macro now.
also, the Editor bit for chapters is now included. the code was messed up. let me know if there is any more changes you need.