Update apa-old-doi-prefix.csl
Thanks for creating this, I am using jabref and am very happy this exists. Just a small update: As far as I have seen, in apa 6th, volume should not be italic and issue number should only be mentioned when page numbers start at 0 with every issue, which only very few journals have. So it would be less work to have to add it in those cases, than having to remove it for all.