Update journal-of-biogeography.csl
From the homepage of the journal, we can get that "When a work has two authors, cite both names every time the reference occurs in text. When a work has three, four, or five authors, cite all authors the first time the reference occurs; subsequent citations include only the surname of the first author followed by et al., (not Italicized and with a period after “al.”) and the year if it is the first citation of the reference within a paragraph.", "Please include up to seven authors in the list (use “&” before last author name). For eight or more authors please list the first six and then use ellipses followed by last author (do not use “&” before last author name)", and "The basic reference form for a journal paper is: Author (date).Paper title. Journal, Volume, page; and for a book citation: Author (date). Book title. Place of publication, publisher.". These features cannot be achieved by the old version, so I propose this version.