Improve edition macro for din-1505-2
Almost any edition i encountered is either plain numeric or - if not - already contains some (abbreviated) form of the word "edition". So in most cases my German citations have edition statements ending in "... Aufl. Aufl.", with the latter occurrence provided by the <text/> element in the style.
Unfortunately, I have no statistical data to prove that this solution is better. 
But at least the modifications in this patch request are based on how the "edition" is formatted in vancouver.csl.