universite-de-liege-droit.csl
We made some improvements, especially concerning the reports (in the first version some variables were mentioned twice). We have also changed the way we refer to Internet URLs and the viewing date of web pages.
For chapters, the formatting of the "in" preceding the title of the work is now conditional: italics when there is an editor or director, roman type when there is neither editor nor director.
We have added support for references to case law, bill and legislation in the Belgian citation style.
Finally, we have reversed the initial and name of the editor and/or director in the bibliography.