Fixed comma setting for books (#3709)
* Fixed another problem with comma setting

Example: If we specified a book with an edition, we had a prefixing comma with whitespace between the prepending element and the book. We may run into other problems involving strange design choices concerning prefixes, but this is an atomic fix for now.

* Fix of book volume

The book volume would have been display as `Vol. , 3` (thanks to @adam3smith to clear that up https://github.com/Braintelligence/styles/commit/4452beb5234952923f8b49169344efeff6ab7ea9#r30674902 😸 ).
This should fix it and make it look like `Vol. 3` instead, so it looks like `Bd. 3` in German as intended.