Create skripsi-sejarah-universitas-gadjah-mada.csl
Hi there,

I've submitted  a citation and bibliography style match for an undergraduate thesis in the History Department, Faculty of Cultural Science, Gadjah Mada University, hence the name. I hope it can help anyone out there, struggling with citation management in Indonesian language.

This style is based on Chicago Full Note with addition/alteration on outdated style ibid., loc. cit., etc. This style also partly translated into Indonesian.

I've done this because, as far as I know, there is none of Indonesian student/researcher really done it. I translated some of the terms used in the citation style. The style used in History department is indeed old and outdated, with old style ibid and his friends, but it is the way it is. So, anyhow, I need to do this because the citation management is crucial in part of writing an undergrad thesis. It cut my time of writing to almost half. 

---
Kind regards,
Muhammad Asyrafi